+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of lizard-nxt client
 Unreleased (2.6.2) (XXXX-XX-XX)
 -------------------------------
 
+- Add default color and order to timeseries.
+
 - Enable toggling timeseries and temporal raster data off in db.
 
 - Add subtle grid in db.